+++ src/Discord/Rest/User.hs
@@ -7,13 +7,20 @@
 -- | Provides actions for Channel API interactions
 module Discord.Rest.User
   ( UserRequest(..)
+  , parseCurrentUserAvatar
+  , CurrentUserAvatar
   ) where
 
 
 import Data.Aeson
+import Codec.Picture
 import Data.Monoid (mempty, (<>))
 import Network.HTTP.Req ((/:))
 import qualified Network.HTTP.Req as R
+import qualified Data.Text as T
+import qualified Data.ByteString.Char8 as Q
+import qualified Data.ByteString.Lazy.Char8 as QL
+import qualified Data.ByteString.Base64 as B64
 
 import Discord.Rest.Prelude
 import Discord.Types
@@ -31,8 +38,8 @@ data UserRequest a where
   GetCurrentUser       :: UserRequest User
   -- | Returns a 'User' for a given user ID
   GetUser              :: Snowflake -> UserRequest User
-  -- | Modify the requestors user account settings. Returns a 'User' object on success.
-  -- todo ModifyCurrentUser    :: ToJSON o => o -> UserRequest User
+  -- | Modify user's username & avatar pic
+  ModifyCurrentUser    :: T.Text -> CurrentUserAvatar -> UserRequest User
   -- | Returns a list of user 'Guild' objects the current user is a member of.
   --   Requires the guilds OAuth2 scope.
   GetCurrentUserGuilds :: UserRequest [PartialGuild]
@@ -43,12 +50,22 @@ data UserRequest a where
   -- | Create a new DM channel with a user. Returns a DM 'Channel' object.
   CreateDM             :: Snowflake -> UserRequest Channel
 
+-- | Formatted avatar data https://discordapp.com/developers/docs/resources/user#avatar-data
+data CurrentUserAvatar = CurrentUserAvatar String
+
+parseCurrentUserAvatar :: Q.ByteString -> Either String CurrentUserAvatar
+parseCurrentUserAvatar bs =
+  case decodeImage bs of
+    Left e -> Left e
+    Right im -> Right $ CurrentUserAvatar $ "data:image/png;base64,"
+                <> Q.unpack (B64.encode (QL.toStrict (encodePng (convertRGBA8 im))))
+
 
 userMajorRoute :: UserRequest a -> String
 userMajorRoute c = case c of
   (GetCurrentUser) ->                        "me "
   (GetUser _) ->                           "user "
-  -- (ModifyCurrentUser _) ->          "modify_user "
+  (ModifyCurrentUser _ _) ->        "modify_user "
   (GetCurrentUserGuilds) ->     "get_user_guilds "
   (LeaveGuild g) ->                 "leave_guild " <> show g
   (GetUserDMs) ->                       "get_dms "
@@ -68,8 +85,9 @@ userJsonRequest c = case c of
 
   (GetUser user) -> Get (users // user ) mempty
 
-  -- (ModifyCurrentUser patch) ->
-      -- Patch (users /: "@me")  (R.ReqBodyJson patch) mempty
+  (ModifyCurrentUser name (CurrentUserAvatar im)) ->
+      Patch (users /: "@me")  (R.ReqBodyJson (object [ "username" .= name
+                                                     , "avatar" .= im ])) mempty
 
   (GetCurrentUserGuilds) -> Get (users /: "@me" /: "guilds") mempty