+++ src/Discord/Rest/Prelude.hs
@@ -24,7 +24,7 @@ authHeader auth =
   where
   -- | https://discordapp.com/developers/docs/reference#user-agent
   -- Second place where the library version is noted
-  agent = "DiscordBot (https://github.com/aquarial/discord-haskell, 0.5.1)"
+  agent = "DiscordBot (https://github.com/aquarial/discord-haskell, 0.5.2)"
 
 -- Append to an URL
 infixl 5 //